+++ Aula_MVC.Domain/Model/VendaItem.cs
@@ -10,23 +10,15 @@ namespace Aula_MVC.Domain.Model
 {
     public class VendaItem : IEntity
     {
+       
         public virtual long Id { get; set; }
         public virtual Produto Produto { get; set; }
         public virtual Venda Venda { get; set; }
 
         public virtual decimal Preco { get; set; }
-        public virtual int Quantidade { get; set; }
+        public virtual decimal Quantidade { get; set; }
         public virtual decimal Total { get; set; }
 
-        //protected VendaItem(Produto produto, int quantidade)
-        //{
-        //    DomainException.When(produto == null, "Produto Obrigatório");
-        //    DomainException.When(quantidade < 1, "Quantidade Incorreta");
-
-        //    Produto = produto;
-        //    Preco = produto.Preco;
-        //    Quantidade = quantidade;
-        //    Total = Preco * Quantidade;
-        //}
+       
     }
 }
